pyproject.toml
[project]
name = "travel-chronicle-pipeline"
version = "0.1.0"
description = "AI-powered audio analysis pipeline for Travel Chronicle using Gemini 3 Flash"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydub>=0.25.0",
    "audioop-lts>=0.2.0; python_version>='3.13'",  # Required for pydub on Python 3.13+
]

[project.urls]
Homepage = "https://github.com/anne-keksi-ai/travel-chronicle-pipeline"
Repository = "https://github.com/anne-keksi-ai/travel-chronicle-pipeline"

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "pytest-timeout>=2.2.0",
    "mypy>=1.8.0",
    "bandit>=1.7.0",
    "pre-commit>=3.6.0",
    "pip-audit>=2.7.0",
]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=.",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
timeout = 300  # 5 minute default timeout for all tests

[tool.mypy]
python_version = "3.9"
exclude = ["^.venv/", "^venv/"]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Start lenient, can make stricter later
disallow_any_unimported = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true
strict_optional = true

# Ignore missing imports for third-party libraries without stubs
[[tool.mypy.overrides]]
module = [
    "google.*",
    "dotenv.*",
    "pydub",
    "pydub.*",
]
ignore_missing_imports = true

# Skip checking pytest internals (causes issues with Python 3.9 compatibility)
[[tool.mypy.overrides]]
module = ["_pytest.*", "pytest.*"]
follow_imports = "skip"

# Allow conftest imports in test files
[[tool.mypy.overrides]]
module = ["conftest"]
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = [
    "tests",
    ".venv",
    "venv",
]
skips = [
    "B101",  # assert_used - we use asserts in tests
]

[tool.bandit.assert_used]
skips = ["*/test_*.py", "tests/*"]
